updated boostrap script to support raspian (stretch), removed dependency on bash, ensuring the script is run as root, and giving the option to install node 7 instead of node 6

+++ bootstrap/bootstrap.sh
@@ -1,4 +1,11 @@
-#!/bin/bash
+#!/bin/sh
+
+#----make sure this is run as root
+user=`id -u`
+if [ $user -ne 0 ]; then
+    echo "This script requires root permissions. Please run this script with sudo."
+    exit
+fi
 
 #----ascii art!
 echo " _   _ _           _     _                 _       _                   "
@@ -24,31 +31,30 @@ echo "packages, removing old packages, etc.)"
 echo "-----------------------------------------------------------------------"
 
 #----confirm bootstrap
-read -p "Would you like to use this Raspberry Pi for TJBot? Y/n: " choice
-shopt -s nocasematch
+read -p "Would you like to use this Raspberry Pi for TJBot? [Y/n] " choice
 case "$choice" in
- "n" )
-    echo "OK, TJBot software will not be installed at this time."
-    exit
-    ;;
- *) ;;
+    "n" | "N")
+        echo "OK, TJBot software will not be installed at this time."
+        exit
+        ;;
+    *) ;;
 esac
 
 #----test raspbian version: 8 is jesse, 9 is stretch
-RASPIAN_VERSION=`cat /etc/os-release | grep VERSION_ID | cut -d '"' -f 2`
-if [ $RASPIAN_VERSION -ne 8 ] && [ $RASPIAN_VERSION -ne 9 ]; then
+RASPIAN_VERSION_ID=`cat /etc/os-release | grep VERSION_ID | cut -d '"' -f 2`
+RASPIAN_VERSION=`cat /etc/os-release | grep VERSION | grep -v ID | cut -d '"' -f 2`
+if [ $RASPIAN_VERSION_ID -ne 8 ] && [ $RASPIAN_VERSION_ID -ne 9 ]; then
     echo "Warning: it looks like your Raspberry Pi is not running Raspian (Jesse)"
     echo "or Raspian (Stretch). TJBot has only been tested on these versions of"
     echo "Raspian."
     echo ""
-    read -p "Would you like to continue with setup? Y/n: " choice
-    shopt -s nocasematch
+    read -p "Would you like to continue with setup? [Y/n] " choice
     case "$choice" in
-        "n" )
+        "n" | "N")
             echo "OK, TJBot software will not be installed at this time."
             exit
             ;;
-        * ) ;;
+        *) ;;
     esac
 fi
 
@@ -58,8 +64,8 @@ echo ""
 echo "Please enter a name for your TJBot. This will be used for the hostname of"
 echo "your Raspberry Pi."
 read -p "TJBot name (current: $CURRENT_HOSTNAME): " name
-shopt -s nocasematch
-if [ -z "${name// }" ]; then
+name=$(echo "$name" | tr -d ' ')
+if [ -z $name ]; then
     name=$CURRENT_HOSTNAME
 fi
 echo "Setting DNS hostname to $name"
@@ -70,87 +76,110 @@ sed -i "s/127.0.1.1.*$CURRENT_HOSTNAME/127.0.1.1\t$name/g" /etc/hosts
 echo ""
 echo "In some networking environments, disabling ipv6 may help your Pi get on"
 echo "the network."
-read -p "Disable ipv6? (y/N): " choice
-shopt -s nocasematch
+read -p "Disable ipv6? [y/N] " choice
 case "$choice" in
- "y" )
-    echo "Disabling ipv6"
-    echo " ipv6.disable=1" | tee -a /boot/cmdline.txt
-    echo "ipv6 has been disabled. It will take effect after rebooting.";;
- *) ;;
+    "y" | "Y")
+        echo "Disabling ipv6"
+        echo " ipv6.disable=1" | tee -a /boot/cmdline.txt
+        echo "ipv6 has been disabled. It will take effect after rebooting."
+        ;;
+    *) ;;
 esac
 
 #----setting DNS to Google
 echo ""
 echo "In some networking environments, using Google's nameservers may speed up"
 echo "DNS queries."
-read -p "Enable Google DNS? (y/N): " choice
-shopt -s nocasematch
+read -p "Enable Google DNS? [y/N]: " choice
 case "$choice" in
- "y" )
-    echo "Adding Google DNS servers to /etc/resolv.conf"
-    if ! grep -q "nameserver 8.8.8.8" /etc/resolv.conf; then
-        echo "nameserver 8.8.8.8" | tee -a /etc/resolv.conf
-        echo "nameserver 8.8.4.4" | tee -a /etc/resolv.conf
-    fi ;;
- *) ;;
+    "y" | "Y")
+        echo "Adding Google DNS servers to /etc/resolv.conf"
+        if ! grep -q "nameserver 8.8.8.8" /etc/resolv.conf; then
+            echo "nameserver 8.8.8.8" | tee -a /etc/resolv.conf
+            echo "nameserver 8.8.4.4" | tee -a /etc/resolv.conf
+        fi
+        ;;
+    *) ;;
 esac
 
 #----setting local to US
 echo ""
-read -p "Force locale to US English (en-US)? (y/N): " choice
-shopt -s nocasematch
+read -p "Force locale to US English (en-US)? [y/N] " choice
 case "$choice" in
- "y" )
-    echo "Forcing locale to en-US. Please ignore any errors below."
-    export LC_ALL="en_US.UTF-8"
-    echo "en_US.UTF-8 UTF-8" | tee -a /etc/locale.gen
-    locale-gen en_US.UTF-8
-    ;;
- *) ;;
+    "y" | "Y")
+        echo "Forcing locale to en-US. Please ignore any errors below."
+        export LC_ALL="en_US.UTF-8"
+        echo "en_US.UTF-8 UTF-8" | tee -a /etc/locale.gen
+        locale-gen en_US.UTF-8
+        ;;
+    *) ;;
 esac
 
 #----update raspberry
 echo ""
 echo "TJBot requires an up-to-date installation of your Raspberry Pi's operating"
 echo "system software."
-read -p "Proceed with apt-get dist-upgrade? (Y/n): " choice
-shopt -s nocasematch
+read -p "Proceed with apt-get dist-upgrade? [Y/n] " choice
 case "$choice" in
- "n" )
-    echo "Warning: you may encounter problems running TJBot recipes without performing"
-    echo "an apt-get dist-upgrade. If you experience these problems, please re-run"
-    echo "the bootstrap script and perform this step."
-    ;;
- *)
-    echo "Updating apt repositories [apt-get update]"
-    apt-get update
-    echo "Upgrading OS distribution [apt-get dist-upgrade]"
-    apt-get -y dist-upgrade
-    ;;
+    "n" | "N")
+        echo "Warning: you may encounter problems running TJBot recipes without performing"
+        echo "an apt-get dist-upgrade. If you experience these problems, please re-run"
+        echo "the bootstrap script and perform this step."
+        ;;
+    *)
+        echo "Updating apt repositories [apt-get update]"
+        apt-get update
+        echo "Upgrading OS distribution [apt-get dist-upgrade]"
+        apt-get -y dist-upgrade
+        ;;
 esac
 
 #----nodejs install
-node_version=$(node --version 2>&1)
+NODE_VERSION=$(node --version 2>&1)
+NODE_LEVEL=$(node --version 2>&1 | cut -d '.' -f 1 | awk '{print substr($0,2,1)}')
+
+# Node.js version 6 for Jesse
+if [ $RASPIAN_VERSION_ID -eq 8 ]; then
+    RECOMMENDED_NODE_LEVEL="6"
+# Node.js version 7 for Stretch
+elif [ $RASPIAN_VERSION_ID -eq 9 ]; then
+    RECOMMENDED_NODE_LEVEL="7"
+# Node.js version 7 for anything else
+else
+    RECOMMENDED_NODE_LEVEL="7"
+fi
+
 echo ""
-echo "TJBot requires Node.js version 6. We detected Node.js version $node_version"
-echo "is already installed."
-read -p "Install Node 6.x? (Y/n): " choice
-shopt -s nocasematch
-case "$choice" in
- "y" )
-    curl -sL https://deb.nodesource.com/setup_6.x | bash -
-    apt-get install -y nodejs
-    ;;
- *)
-    echo "Warning: TJBot will encounter problems with versions of Node.js older than 6.x."
-    echo "TJBot has not been tested with Node.js version 7 or higher."
-    ;;
-esac
+if [ $NODE_LEVEL -ge $RECOMMENDED_NODE_LEVEL ]; then
+    echo "Node.js version $NODE_VERSION is installed, which is the recommended version for"
+    echo "Raspian $RASPIAN_VERSION. Congratulations!"
+else
+    echo "Node.js version $NODE_VERSION is currently installed. We recommend installing"
+    echo "Node.js version $RECOMMENDED_NODE_LEVEL for Raspian $RASPIAN_VERSION."
+
+    read -p "Would you like to install a newer version of Node.js? [Y/n] " choice
+    case "$choice" in
+        "y" | "Y")
+            read -p "Which version of Node.js would you like to install? [6/7] " node_version
+            case "$node_version" in
+                "6" | "7")
+                    curl -sL https://deb.nodesource.com/setup_${node_version}.x | sudo bash -
+                    apt-get install -y nodejs
+                    ;;
+                *)
+                    echo "Invalid Node.js version specified, Node.js will not be upgraded at this time."
+                    ;;
+            esac
+            ;;
+        *)
+            echo "Warning: TJBot will encounter problems with versions of Node.js older than 6.x."
+            ;;
+    esac
+fi
 
 #----install additional packages
 echo ""
-if [ $RASPIAN_VERSION -eq 8 ]; then
+if [ $RASPIAN_VERSION_ID -eq 8 ]; then
     echo "Installing additional software packages for Jesse (alsa, libasound2-dev, git, pigpio)"
     apt-get install -y alsa-base alsa-utils libasound2-dev git pigpio
 #elif [ $RASPIAN_VERSION -eq 9 ]; then
@@ -166,37 +195,36 @@ apt-get -y autoremove
 #----enable camera on raspbery pi
 echo ""
 echo "If your Raspberry Pi has a camera installed, TJBot can use it to see."
-read -p "Enable camera? (y/N): " choice
-shopt -s nocasematch
+read -p "Enable camera? [y/N] " choice
 case "$choice" in
- "y" )
-    if grep "start_x=1" /boot/config.txt
-    then
-        echo "Camera is alredy enabled."
-    else
-        echo "Enabling camera."
-        if grep "start_x=0" /boot/config.txt
+    "y" | "Y")
+        if grep "start_x=1" /boot/config.txt
         then
-            sed -i "s/start_x=0/start_x=1/g" /boot/config.txt
+            echo "Camera is alredy enabled."
         else
-            echo "start_x=1" | tee -a /boot/config.txt >/dev/null 2>&1
+            echo "Enabling camera."
+            if grep "start_x=0" /boot/config.txt
+            then
+                sed -i "s/start_x=0/start_x=1/g" /boot/config.txt
+            else
+                echo "start_x=1" | tee -a /boot/config.txt >/dev/null 2>&1
+            fi
+            if grep "gpu_mem=128" /boot/config.txt
+            then
+                :
+            else
+                echo "gpu_mem=128" | tee -a /boot/config.txt >/dev/null 2>&1
+            fi
         fi
-        if grep "gpu_mem=128" /boot/config.txt
-        then
-            :
-        else
-            echo "gpu_mem=128" | tee -a /boot/config.txt >/dev/null 2>&1
-        fi
-    fi
-    ;;
- *) ;;
+        ;;
+    *) ;;
 esac
 
 #----clone tjbot
 echo ""
 echo "We are ready to clone the TJBot project."
 read -p "Where should we clone it to? (default: /home/pi/Desktop/tjbot): " TJBOT_DIR
-if [ -z "${TJBOT_DIR// }" ]; then
+if [ -z $TJBOT_DIR ]; then
     TJBOT_DIR='/home/pi/Desktop/tjbot'
 fi
 
@@ -215,22 +243,23 @@ echo "speaker via HDMI, USB, or Bluetooth, this is a safe operation and you will
 echo "be able to play sound and use the LED at the same time. If you plan to use"
 echo "the built-in audio jack, we recommend NOT disabling the sound kernel"
 echo "modules."
-read -p "Disable sound kernel modules? (Y/n): " choice
-shopt -s nocasematch
+read -p "Disable sound kernel modules? [Y/n] " choice
 case "$choice" in
- "y" )
-    echo "Disabling the kernel modules for the built-in audio jack."
-    cp $TJBOT_DIR/bootstrap/tjbot-blacklist-snd.conf /etc/modprobe.d/
-    ;;
- "n" )
-    echo "Enabling the kernel modules for the built-in audio jack."
-    rm /etc/modprobe.d/tjbot-blacklist-snd.conf
-    ;;
- *) ;;
+    "y" | "Y")
+        echo "Disabling the kernel modules for the built-in audio jack."
+        cp $TJBOT_DIR/bootstrap/tjbot-blacklist-snd.conf /etc/modprobe.d/
+        ;;
+    "n" | "N")
+        if [ -f /etc/modprobe.d/tjbot-blacklist-snd.conf ]; then
+            echo "Enabling the kernel modules for the built-in audio jack."
+            rm /etc/modprobe.d/tjbot-blacklist-snd.conf
+        fi
+        ;;
+    *) ;;
 esac
 
 #----installation complete
-sleep_time=0.2
+sleep_time=0.1
 echo ""
 echo ""
 echo "                           .yNNs\`                           "
@@ -301,7 +330,7 @@ echo "Setup complete. Your Raspberry Pi is now set up as a TJBot! ;)"
 sleep $sleep_time
 echo "-------------------------------------------------------------------"
 echo ""
-read -p "Press enter to continue: "
+read -p "Press enter to continue" nonce
 
 #——instructions for watson credentials
 echo "Notice about Watson services: Before running any recipes, you will need"
@@ -324,7 +353,7 @@ these in the config.js files for each recipe you wish to run."
 echo "For more detailed guides on setting up service credentials, please see the
 README file of each recipe, or search instructables.com for \"tjbot\"."
 echo ""
-read -p "Press enter to continue: "
+read -p "Press enter to continue" nonce
 
 #----tests
 echo ""
@@ -333,10 +362,9 @@ echo "functioning properly. If you have made any changes to the camera or"
 echo "sound configuration, we recommend rebooting first before running these"
 echo "tests as they may fail. You can run these tests at anytime by running"
 echo "the runTests.sh script in the tjbot/bootstrap folder."
-read -p "Would you like to run hardware tests at this time? (y/N): " choice
-shopt -s nocasematch
+read -p "Would you like to run hardware tests at this time? [y/N] " choice
 case "$choice" in
-    "y" )
+    "y" | "Y")
         ./runTests.sh $TJBOT_DIR
         ;;
     *) ;;
@@ -344,10 +372,9 @@ esac
 
 #----reboot
 echo ""
-read -p "We recommend rebooting for all changes to take effect. Reboot? (Y/n): " choice
-shopt -s nocasematch
+read -p "We recommend rebooting for all changes to take effect. Reboot? [Y/n] " choice
 case "$choice" in
-    "y" )
+    "y" | "Y")
         echo "Rebooting."
         reboot
         ;;
